chore: test number of nodes on edition
