+++ ui/packages/types/src/db-type/index.ts
@@ -1,13 +1,15 @@
 enum DbType {
-  Postresql = 'postgresql',
-  Mongo = 'mongodb',
-  Mysql = 'mysql',
+  Postresql = "postgresql",
+  Mongo = "mongodb",
+  Mysql = "mysql",
 }
 
 enum DbEngineType {
-  PSMDB = 'psmdb',
-  PXC = 'pxc',
-  POSTGRESQL = 'postgresql',
+  PSMDB = "psmdb",
+  PXC = "pxc",
+  POSTGRESQL = "postgresql",
 }
+
+export type ProxyType = "mongos" | "haproxy" | "proxysql" | "pgbouncer";
 
 export { DbType, DbEngineType };
